Only remove user_attributes from the table if it exists

+++ db/migrate/20171127214039_remove_user_attributes_from_users.rb
@@ -1,6 +1,6 @@
 class RemoveUserAttributesFromUsers < ActiveRecord::Migration
   def up
-    remove_column :users, :user_attributes
+    remove_column :users, :user_attributes if column_exists? :users, :user_attributes
   end
 
   def down
